Move tests back onto parallel directory.
