+++ example/spec/parallel/ingestGranule/IngestGranule.UMM.output.payload.template.json
